exec: Add node tampering blocklist function
We'll be adding checks to block writes from processes which tamper with
values that we control from within the kernel, especially ones that
userspace writes to for boosting. Add a central function to perform the
process check to reduce code duplication.

This blocklists the following processes which are known to tamper with
such values:
  - init
  - libperfmgr (power-servic and NodeLooperThrea)
  - perfd (PERFD-SERVER)
  - init.qcom.post_boot.sh (init.qcom.post_)
  - Processes with names starting with: iop@, perf@, power@

Change-Id: I2b0736c2c42d575e46654b1071b669b8959d7c93
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Pzqqt <821026875@qq.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>
Signed-off-by: kylieeXD <khayloaftck@gmail.com>